+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.2 (2023-06-08)
 
 - Added `Card` component along with `CardBody`, `CardActions`, and `CardTitle` sub components.
+- Added `Alert` component.
 
 ## 1.0.1 (2023-06-08)